Kill the JVM if the Kafka consumer stage ever stops
This only occurs when something really bad and the Kafka consumer actor
dies. We can't recover from it, so kill the JVM and let ECS replace the
Docker container.

Also get rid of the Kafka actor watching stuff. It worked, but looking
up an actor by name is a bit dodgy. It was only a temporary debug thing.